pyproject.toml
[project]
name = "agent-development"
version = "0.1.0"
description = "A LangChain agent development project"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.1.0",
    "langchain-ollama>=1.0.1",
    "langchain-tavily>=0.2.15",
    "python-dotenv>=1.2.1",
]

